Replaced: package list -> find_packages()

+++ setup.py
@@ -8,12 +8,12 @@
 # GIT: https://github.com/IsmaelRLG/simpbot
 
 try:
-    from setuptools import setup
+    from setuptools import setup, find_packages
 except ImportError:
     print("Please install the setuptools package")
     exit(0)
 
-from os.path import join
+from os.path import join, exists
 from os.path import dirname as dir
 import simpbot
 import platform
@@ -23,7 +23,8 @@ currentdir = dir(__file__)
 bindir = join(currentdir, 'bin')
 script = 'simpbot.py'
 if platform.system() == 'Linux':
-    shutil.move(join(bindir, script), join(bindir, 'simpbot'))
+    if exists(join(bindir, script)):
+        shutil.move(join(bindir, script), join(bindir, 'simpbot'))
     script = 'simpbot'
 
 requires = file(join(currentdir, 'requires.txt'), 'r').read().splitlines()
@@ -35,19 +36,7 @@ setup(
     author_email="ismaelrlgv@gmail.com",
     description="Simple Bot (SimpBot) - IRC (Internet Relay Chat) Bot",
     url="https://github.com/IsmaelRLG/simpbot",
-    packages=[
-        'simpbot',
-        'simpbot.request',
-        'simpbot.admins',
-        'simpbot.dbstore',
-        'simpbot.bottools',
-        'simpbot.moduletools',
-        'simpbot.samples',
-        'simpbot.commands',
-        'simpmods',
-        'simpmods.admin'
-        ],
-
+    packages=find_packages(),
     package_data={'simpbot': [join('localedata', '*.dat')]},
     scripts=[
         join('bin', script)
@@ -59,8 +48,7 @@ setup(
         "Development Status :: 4 - Beta",
         "Environment :: Console",
         "License :: OSI Approved :: MIT License",
-        "Natural Language :: Spanish",
-        "Operating System :: POSIX",
+        "Operating System :: POSIX :: Linux",
         "Operating System :: Microsoft :: Windows",
         "Programming Language :: Python :: 2.7",
         'Programming Language :: Python :: 3',
